refactor: 상속 구조를 고쳐서 메서드의 중복 라인 제거
