+++ src/main/java/chess/domain/piece/common/Rook.java
@@ -5,7 +5,6 @@ import static chess.domain.piece.Direction.N;
 import static chess.domain.piece.Direction.S;
 import static chess.domain.piece.Direction.W;
 
-import chess.domain.board.Position;
 import chess.domain.piece.Color;
 import chess.domain.piece.Direction;
 import chess.domain.piece.PieceType;
@@ -22,15 +21,6 @@ public class Rook extends CommonPiece {
     @Override
     public boolean isEmpty() {
         return false;
-    }
-
-    @Override
-    public Direction findValidDirection(final Position current, final Position target) {
-        final int columnDifference = target.calculateColumnDifference(current);
-        final int rowDifference = target.calculateRowDifference(current);
-        final Direction direction = Direction.calculate(columnDifference, rowDifference);
-        validateDirection(direction);
-        return direction;
     }
 
     @Override
