+++ src/main/java/chess/domain/piece/ranged/RangedPiece.java
@@ -3,31 +3,13 @@ package chess.domain.piece.ranged;
 import chess.domain.board.Position;
 import chess.domain.piece.Color;
 import chess.domain.piece.Direction;
-import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import java.util.ArrayList;
-import java.util.List;
+import chess.domain.piece.common.CommonPiece;
 
-public abstract class RangedPiece extends Piece {
+public abstract class RangedPiece extends CommonPiece {
 
     protected RangedPiece(PieceType pieceType, Color color) {
         super(pieceType, color);
-    }
-
-    @Override
-    public List<Position> calculatePathToValidate(final Position current, final Position target,
-                                                  final Piece targetPiece) {
-        Direction direction = findValidDirection(current, target);
-        List<Position> path = new ArrayList<>();
-        Position moved = current.move(direction);
-        while (!moved.equals(target)) {
-            path.add(moved);
-            moved = moved.move(direction);
-        }
-        if (hasSameColor(targetPiece)) {
-            path.add(target);
-        }
-        return path;
     }
 
     @Override
@@ -43,9 +25,4 @@ public abstract class RangedPiece extends Piece {
     protected abstract void validateDirection(Direction direction);
 
     protected abstract void validateRange(int columnDifference, int rowDifference);
-
-    @Override
-    public boolean isEmpty() {
-        return false;
-    }
 }
